Day 9, A solved: Integer wasn't enough :D

+++ 9/ab.scala
@@ -58,8 +58,5 @@ def a(input: String, debug: Boolean = false) =
     println
   
   getRearrangedBlocks(blocks, freeIndexes, debug).zipWithIndex
-    .map((maybeInt, i) => maybeInt.getOrElse(0) * i)
+    .map((maybeInt, i) => maybeInt.getOrElse(0).toLong * i.toLong)
     .sum
-
-// Wrong answers
-// 829320760
